backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0

# Google Gemini AI
google-generativeai==0.3.2

# Vector Search (optional but recommended)
faiss-cpu==1.7.4
numpy==1.24.3

# Environment variables
python-dotenv==1.0.0

# Optional: for better token counting
# tiktoken==0.5.1

# Optional: for OCR fallback
# pytesseract==0.3.10
# pdf2image==1.16.3
# Pillow==10.1.0